+++ MRSTWEb.Domain/Repositories/DiscountRepository.cs
@@ -3,36 +3,68 @@ using MRSTWEb.Domain.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Caching;
 
 namespace MRSTWEb.Domain.Repositories
 {
     public class DiscountRepository : IRepository<Discount>
     {
         private EF.AppContext db;
-        public DiscountRepository(EF.AppContext db)
+        private MemoryCache _cache = MemoryCache.Default;
+        public DiscountRepository()
         {
-            this.db = db;
+            this.db = new EF.AppContext();
         }
         public void Create(Discount item)
         {
             db.Discounts.Add(item);
-
+            db.SaveChanges();
+            InvalidateCache();
         }
 
         public void Delete(int id)
         {
             var discount = db.Discounts.Find(id);
-            db.Discounts.Remove(discount);
+            if (discount != null)
+            {
+                db.Discounts.Remove(discount);
+                db.SaveChanges();
+                InvalidateCache();
+            }
         }
 
         public Discount Get(int id)
         {
-            return db.Discounts.Find(id);
+            string cacheKey = $"Discount_{id}";
+            if (_cache.Contains(cacheKey))
+            {
+                return (Discount)_cache.Get(cacheKey);
+            }
+
+            var discount = db.Discounts.Find(id);
+            if (discount != null)
+            {
+                _cache.Set(cacheKey, discount, DateTimeOffset.Now.AddMinutes(10));
+            }
+
+            return discount;
         }
 
         public IEnumerable<Discount> GetAll()
         {
-            return db.Discounts.ToList();
+            string cacheKey = "AllDiscounts";
+            if (_cache.Contains(cacheKey))
+            {
+                return (IEnumerable<Discount>)_cache.Get(cacheKey);
+            }
+
+            var discounts = db.Discounts.ToList();
+            if (discounts != null && discounts.Any())
+            {
+                _cache.Set(cacheKey, discounts, DateTimeOffset.Now.AddMinutes(10));
+            }
+
+            return discounts;
         }
 
         public IEnumerable<Order> GetAllOrdersWithUsers(string userId)
@@ -58,6 +90,22 @@ namespace MRSTWEb.Domain.Repositories
                 discount.ExpirationTime = item.ExpirationTime;
                 discount.SetTime = item.SetTime;
                 discount.Percentage = item.Percentage;
+                db.Entry(discount).State = System.Data.Entity.EntityState.Modified;
+                db.SaveChanges();
+                InvalidateCache();
+            }
+        }
+
+        public void UpdateBookDiscount(decimal Price, int bookId)
+        {
+            throw new NotImplementedException();
+        }
+
+        private void InvalidateCache()
+        {
+            foreach (var item in _cache)
+            {
+                _cache.Remove(item.Key);
             }
         }
     }
